% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/First functions.R
\name{Find_possible_values}
\alias{Find_possible_values}
\title{Check for unused numbers between 1 and 9}
\usage{
Find_possible_values(Filled_already)
}
\arguments{
\item{Filled_already}{A vector of length 9 from a sudoku puzzle}
}
\value{
Vector of numbers which have not been used
}
\description{
Takes a vector and compares against the numbers 1 to 9, to find which are
missing.
}
\examples{

Find_possible_values(c(1,2,3,4,NA,NA,7,8,9)) # Test a single row or column

}
