% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSudoku.R
\name{plot.sudoku}
\alias{plot.sudoku}
\title{Function to plot the Sudoku.}
\usage{
\method{plot}{sudoku}(sudoku, ...)
}
\arguments{
\item{sudoku}{A Sudoku object}

\item{...}{Further argument passed to or from other methods.}
}
\description{
This function takes an object with the S3 Sudoku class and plots it as a sudoku.
}
