% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveSudoku.R
\name{solveSudoku}
\alias{solveSudoku}
\title{Sudoku Solver}
\usage{
solveSudoku(sudoku)
}
\arguments{
\item{sudoku}{A object with data.frame, matrix or sudoku. It must range Must comply with general sudoku rules so have 9 rows and
columns and have numbers that range from 1 to 9. Empty spots must be denoted with a 0 or an NA. If the input sudoku has an error,
the function can result in random errors.}
}
\value{
Returns the solved sudoku with an s3 class called sudoku. This object can be plotted to get a clear
visualization of the solved sudoku. This function also check if the sollution is correct and automaticly plots the sudoku.
}
\description{
This function solves a sudoku. It does this by using the follow techniques: lone, hidden and naked singles,
hidden and naked pairs(and triples and quadruples), pointing pairs and triples, box line reduction, x-wings (and swordfishes and
jellyfishes), y-wings and skyscrapers. When another technique is needed to solve the sudoku, this function will
not be able to solve the sudoku.
}
