% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/First functions.R
\name{Sudoku_solver_np}
\alias{Sudoku_solver_np}
\title{Solve a sudoku puzzle}
\usage{
Sudoku_solver_np(Puzzle)
}
\arguments{
\item{Puzzle}{A 9 x 9 numeric matrix representing a sudoku puzzle, with NA values for gaps.}
}
\value{
a 9 x 9 matrix with all cells filled.
}
\description{
Takes a 9 X 9 matrix with missing values and replaces this with the solution.
}
