% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudokuHint.R
\name{sudokuHint}
\alias{sudokuHint}
\title{Get a hint to solve the sudoku}
\usage{
sudokuHint(sudoku, randomHint = TRUE)
}
\arguments{
\item{sudoku}{A object with data.frame, matrix or sudoku. It must range Must comply with general sudoku rules so have 9 rows and
columns and have numbers that range from 1 to 9. Empty spots must be denoted with a 0 or an NA. If the input sudoku has an error,
the function can result in random errors.}

\item{randomHint}{If TRUE, the user will get all the hints. If FALSE, the user can select a desired hint. TRUE by default.}
}
\value{
Returns a (not completed) sudoku.
}
\description{
This function gives hints to solve sudokus. It can give hints for the following situations:hidden and naked singles,
hidden and naked pairs(and triples and quadruples), pointing pairs and triples, box line reduction, x-wings (and swordfishes and
jellyfishes), y-wings and skyscrapers. The user can ask for all hints or select a certain technique from above. It will provide the
hints one at the time and asks in the console if the user want to move on.
}
